+++ examples/demo-rollup/tests/bank/zk_rollup/bank_tests.rs
@@ -1,13 +1,12 @@
 use anyhow::Context;
-use demo_stf::genesis_config::GenesisPaths;
 use futures::StreamExt;
 use sov_bank::event::Event as BankEvent;
 use sov_bank::utils::TokenHolder;
 use sov_bank::Coins;
 use sov_cli::NodeClient;
-use sov_kernels::basic::BasicKernelGenesisPaths;
 use sov_mock_da::storable::service::StorableMockDaService;
 use sov_mock_da::BlockProducingConfig;
+use sov_modules_api::OperatingMode;
 use sov_rollup_interface::node::da::DaServiceWithRetries;
 use sov_rollup_interface::node::ledger_api::FinalityStatus;
 use sov_test_utils::TestSpec;
@@ -29,10 +28,7 @@ async fn bank_tx_tests_instant_finality_using_sequencer_tx_submission() -> anyho
         get_appropriate_rollup_prover_config(),
         BLOCK_PRODUCING_CONFIG,
         test_case.finalization_blocks,
-        GenesisPaths::from_dir("../test-data/genesis/integration-tests"),
-        BasicKernelGenesisPaths {
-            chain_state: "../test-data/genesis/integration-tests/chain_state_zk.json".into(),
-        },
+        OperatingMode::Zk,
     )
     .await?;
 
@@ -57,10 +53,7 @@ async fn bank_tx_tests_non_instant_finality_using_sequencer_tx_submission() -> a
         get_appropriate_rollup_prover_config(),
         BLOCK_PRODUCING_CONFIG,
         test_case.finalization_blocks,
-        GenesisPaths::from_dir("../test-data/genesis/integration-tests"),
-        BasicKernelGenesisPaths {
-            chain_state: "../test-data/genesis/integration-tests/chain_state_zk.json".into(),
-        },
+        OperatingMode::Zk,
     )
     .await?;
 
@@ -86,10 +79,7 @@ async fn bank_tx_tests_instant_finality_using_da_layer_tx_submission() -> anyhow
         get_appropriate_rollup_prover_config(),
         BLOCK_PRODUCING_CONFIG,
         test_case.finalization_blocks,
-        GenesisPaths::from_dir("../test-data/genesis/integration-tests"),
-        BasicKernelGenesisPaths {
-            chain_state: "../test-data/genesis/integration-tests/chain_state_zk.json".into(),
-        },
+        OperatingMode::Zk,
     )
     .await?;