+++ examples/demo-rollup/tests/evm/evm_test_helper.rs
@@ -1,20 +1,18 @@
 use std::net::SocketAddr;
 
-use demo_stf::genesis_config::GenesisPaths;
 use ethers_core::abi::Address;
 use ethers_providers::ProviderError;
 use ethers_signers::{LocalWallet, Signer};
 use futures::future::join_all;
 use futures::stream::BoxStream;
 use futures::StreamExt;
-use sov_kernels::basic::BasicKernelGenesisPaths;
 use sov_mock_da::{BlockProducingConfig, MockAddress, MockDaConfig};
 use sov_stf_runner::processes::RollupProverConfig;
 use sov_test_utils::SimpleStorageContract;
 use tokio::task::JoinHandle;
 
 use super::test_client::TestClient;
-use crate::test_helpers::start_rollup_in_background;
+use crate::test_helpers::{start_rollup_in_background, test_genesis_paths};
 
 /// Starts test rollup node.  
 pub(crate) async fn start_node(
@@ -30,10 +28,7 @@ pub(crate) async fn start_node(
 
             rpc_port_tx,
             rest_port_tx,
-            GenesisPaths::from_dir("../test-data/genesis/integration-tests"),
-            BasicKernelGenesisPaths {
-                chain_state: "../test-data/genesis/integration-tests/chain_state_zk.json".into(),
-            },
+            test_genesis_paths(sov_modules_api::OperatingMode::Zk),
             rollup_prover_config,
             MockDaConfig {
                 connection_string: "sqlite::memory:".to_string(),